Added alert if the configuration is not correct

+++ src/wazuh_modules/wmodules.c
@@ -161,7 +161,8 @@ void wm_free(wmodule * config) {
 
     for (cur_module = config; cur_module; cur_module = next_module) {
         next_module = cur_module->next;
-        cur_module->context->destroy(cur_module->data);
+        if (cur_module->context)
+            cur_module->context->destroy(cur_module->data);
         free(cur_module);
     }
 }
